+++ scarletio/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.0.80'
+__version__ = '1.0.81'
 
 
 from .core import *